requirements.txt
fastapi
uvicorn[standard]
python-dotenv
requests
currencyconverter
